requirements.txt
Flask==2.0.3
loguru==0.6.0
PyYAML==6.0
schedule==1.1.0
selenium==4.1.3
webdriver_manager==3.5.3
